MDL-16596 support for total maxbytes per draft area - proper server-side validation still missing; some minor refactoring

+++ files/draftfiles.php
@@ -8,6 +8,7 @@
     $newdirname = optional_param('newdirname', '', PARAM_FILE);
     $delete     = optional_param('delete', '', PARAM_PATH);
     $subdirs    = optional_param('subdirs', 0, PARAM_BOOL);
+    $maxbytes   = optional_param('maxbytes', 0, PARAM_INT);
 
     require_login();
     if (isguestuser()) {
@@ -17,6 +18,8 @@
     if (!$context = get_context_instance(CONTEXT_USER, $USER->id)) {
         print_error('invalidcontext');
     }
+
+    $notice = '';
 
     $contextid = $context->id;
     $filearea  = 'user_draft';
@@ -35,25 +38,38 @@
     $files = $fs->get_directory_files($context->id, 'user_draft', $itemid, $directory->get_filepath());
     $parent = $directory->get_parent_directory();
 
+    $totalbytes = 0;
+    foreach ($files as $hash=>$file) {
+        if (!$subdirs and $file->get_filepath() !== '/') {
+            unset($files[$hash]);
+            continue;
+        }
+        $totalbytes += $file->get_filesize();
+    }
+
 /// process actions
     if ($newdirname !== '' and data_submitted() and confirm_sesskey()) {
         $newdirname = $directory->get_filepath().$newdirname.'/';
         $fs->create_directory($contextid, $filearea, $itemid, $newdirname, $USER->id);
-        redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($newdirname).'&amp;subdirs='.$subdirs);
+        redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($newdirname).'&amp;subdirs='.$subdirs.'&amp;maxbytes='.$maxbytes);
     }
 
     if (isset($_FILES['newfile']) and data_submitted() and confirm_sesskey()) {
-        $file = $_FILES['newfile'];
-        $newfilename = clean_param($file['name'], PARAM_FILE);
-        // TODO: some better error handling or use some upload manager
-        if (is_uploaded_file($_FILES['newfile']['tmp_name'])) {
-            if ($existingfile = $fs->get_file($contextid, $filearea, $itemid, $filepath, $newfilename)) {
-                $existingfile->delete();
+        if (!empty($_FILES['newfile']['error'])) {
+            $notice = file_get_upload_error($_FILES['newfile']['error']);
+        } else {
+            $file = $_FILES['newfile'];
+            $newfilename = clean_param($file['name'], PARAM_FILE);
+            // TODO: some better error handling or use some upload manager
+            if (is_uploaded_file($_FILES['newfile']['tmp_name'])) {
+                if ($existingfile = $fs->get_file($contextid, $filearea, $itemid, $filepath, $newfilename)) {
+                    $existingfile->delete();
+                }
+                $filerecord = array('contextid'=>$contextid, 'filearea'=>$filearea, 'itemid'=>$itemid, 'filepath'=>$filepath,
+                                    'filename'=>$newfilename, 'userid'=>$USER->id);
+                $newfile = $fs->create_file_from_pathname($filerecord, $_FILES['newfile']['tmp_name']);
+                redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($filepath).'&amp;subdirs='.$subdirs.'&amp;maxbytes='.$maxbytes);
             }
-            $filerecord = array('contextid'=>$contextid, 'filearea'=>$filearea, 'itemid'=>$itemid, 'filepath'=>$filepath,
-                                'filename'=>$newfilename, 'userid'=>$USER->id);
-            $newfile = $fs->create_file_from_pathname($filerecord, $_FILES['newfile']['tmp_name']);
-            redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($filepath).'&amp;subdirs='.$subdirs);
         }
     }
 
@@ -71,14 +87,18 @@
             $isdir = $file->is_directory();
             $file->delete();
             if ($isdir) {
-                redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($parent->get_filepath()).'&amp;subdirs='.$subdirs);
+                redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($parent->get_filepath()).'&amp;subdirs='.$subdirs.'&amp;maxbytes='.$maxbytes);
             } else {
-                redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($filepath).'&amp;subdirs='.$subdirs);
+                redirect('draftfiles.php?itemid='.$itemid.'&amp;filepath='.rawurlencode($filepath).'&amp;subdirs='.$subdirs.'&amp;maxbytes='.$maxbytes);
             }
         }
     }
 
     print_header();
+
+    if ($notice !== '') {
+        notify($notice);
+    }
 
     echo '<div class="areafiles">';
 
@@ -89,7 +109,7 @@
 
     if ($parent) {
         echo '<div class="folder">';
-        echo '<a href="draftfiles.php?itemid='.$itemid.'&amp;filepath='.$parent->get_filepath().'&amp;subdirs='.$subdirs.'"><img src="'.$CFG->pixpath.'/f/parent.gif" class="icon" alt="" />&nbsp;'.get_string('parentfolder').'</a>';
+        echo '<a href="draftfiles.php?itemid='.$itemid.'&amp;filepath='.$parent->get_filepath().'&amp;subdirs='.$subdirs.'&amp;maxbytes='.$maxbytes.'"><img src="'.$CFG->pixpath.'/f/parent.gif" class="icon" alt="" />&nbsp;'.get_string('parentfolder').'</a>';
         echo '</div>';
     }
 
@@ -107,8 +127,8 @@
                 $dirname = explode('/', trim($filepath, '/'));
                 $dirname = array_pop($dirname);
                 echo '<div class="folder">';
-                echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;subdirs=$subdirs\"><img src=\"$CFG->pixpath/f/folder.gif\" class=\"icon\" alt=\"$strfolder\" />&nbsp;".s($dirname)."</a> ";
-                echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;delete=$filenameurl&amp;subdirs=$subdirs\"><img src=\"$CFG->pixpath/t/delete.gif\" class=\"iconsmall\" alt=\"$strdelete\" /></a>";
+                echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;subdirs=$subdirs&amp;maxbytes=$maxbytes\"><img src=\"$CFG->pixpath/f/folder.gif\" class=\"icon\" alt=\"$strfolder\" />&nbsp;".s($dirname)."</a> ";
+                echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;delete=$filenameurl&amp;subdirs=$subdirs&amp;maxbytes=$maxbytes\"><img src=\"$CFG->pixpath/t/delete.gif\" class=\"iconsmall\" alt=\"$strdelete\" /></a>";
                 echo '</div>';
             }
 
@@ -117,27 +137,42 @@
             $viewurl = $browser->encodepath("$CFG->wwwroot/draftfile.php", "/$contextid/user_draft/$itemid".$filepath.$filename, false, false);
             echo '<div class="file">';
             echo "<a href=\"$viewurl\"><img src=\"$CFG->pixpath/f/$icon\" class=\"icon\" alt=\"$strfile\" />&nbsp;".s($filename)." ($filesize)</a> ";
-            echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;delete=$filenameurl&amp;subdirs=$subdirs\"><img src=\"$CFG->pixpath/t/delete.gif\" class=\"iconsmall\" alt=\"$strdelete\" /></a>";;
+            echo "<a href=\"draftfiles.php?itemid=$itemid&amp;filepath=$filepath&amp;delete=$filenameurl&amp;subdirs=$subdirs&amp;maxbytes=$maxbytes\"><img src=\"$CFG->pixpath/t/delete.gif\" class=\"iconsmall\" alt=\"$strdelete\" /></a>";;
             echo '</div>';
         }
     }
 
     echo '</div>';
 
-    echo '<form enctype="multipart/form-data" method="post" action="draftfiles.php"><div>';
-    echo '<input type="hidden" name="itemid" value="'.$itemid.'" />';
-    echo '<input type="hidden" name="filepath" value="'.s($filepath).'" />';
-    echo '<input type="hidden" name="subdirs" value="'.$subdirs.'" />';
-    echo '<input type="hidden" name="sesskey" value="'.sesskey().'" />';
-    echo '<input name="newfile" type="file" />';
-    echo '<input type="submit" value="'.get_string('uploadafile').'" />';
-    echo '</div></form>';
+    if ($maxbytes == 0 or $maxbytes > $totalbytes) {
+        echo '<form enctype="multipart/form-data" method="post" action="draftfiles.php"><div>';
+        if ($maxbytes) {
+            echo '<input type="hidden" name="MAX_FILE_SIZE" value="'.($maxbytes-$totalbytes).'" />';
+        }
+        echo '<input type="hidden" name="itemid" value="'.$itemid.'" />';
+        echo '<input type="hidden" name="filepath" value="'.s($filepath).'" />';
+        echo '<input type="hidden" name="subdirs" value="'.$subdirs.'" />';
+        echo '<input type="hidden" name="maxbytes" value="'.$maxbytes.'" />';
+        echo '<input type="hidden" name="sesskey" value="'.sesskey().'" />';
+        echo '<input name="newfile" type="file" />';
+        echo '<input type="submit" value="'.get_string('uploadafile').'" />';
+        if ($maxbytes) {
+            echo ' ('.get_string('maxsize', '', display_size(get_max_upload_file_size($CFG->maxbytes, $maxbytes-$totalbytes))).')';
+        } else {
+            echo ' ('.get_string('maxsize', '', display_size(get_max_upload_file_size($CFG->maxbytes))).')';
+        }
+        echo '</div></form>';
+    } else {
+        //TODO: notify upload limit reached here
+        echo get_string('maxsize', '', display_size(get_max_upload_file_size($CFG->maxbytes, $maxbytes)));
+    }
 
     if ($subdirs) {
         echo '<form action="draftfiles.php" method="post"><div>';
         echo '<input type="hidden" name="itemid" value="'.$itemid.'" />';
         echo '<input type="hidden" name="filepath" value="'.s($filepath).'" />';
         echo '<input type="hidden" name="subdirs" value="'.$subdirs.'" />';
+        echo '<input type="hidden" name="maxbytes" value="'.$maxbytes.'" />';
         echo '<input type="hidden" name="sesskey" value="'.sesskey().'" />';
         echo '<input type="text" name="newdirname" value="" />';
         echo '<input type="submit" value="'.get_string('makeafolder').'" />';
